requirements.txt
pygame==2.5.2
numpy==1.24.3
pynput==1.7.6
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7 